ts cleanup, remove optionals
